+++ dataset.h
@@ -7,7 +7,7 @@ struct Sample {
     // All numeric measurements of one flower.
     // For Iris this means:
     // sepal length, sepal width, petal length, petal width
-    std::vector<double> features;
+    std::vector<float> features;
 
     // Correct class/species of this flower.
     std::string label;
